Extract DSS Android bytecode patch into reusable gradle script (dss-android-patch.gradle.kts) and update documentation for integration clarity.
